changed function to operate with double type operands

+++ src/Algorithm.cpp
@@ -5,9 +5,9 @@
  *
  * @param num_1 First operand.
  * @param num_2 Second operand.
- * @return int Result of `num_1` + `num_2`.
+ * @return double Result of `num_1` + `num_2`.
  */
-int AddTwoNumbers(int num_1, int num_2)
+double AddTwoNumbers(double num_1, double num_2)
 {
 
 	return num_1 + num_2;
