+++ src/main.cpp
@@ -1,21 +1,21 @@
-#include<iostream>
-#include "Algorithm.h"
+#include "cpp-template/Algorithm.h"
+#include <iostream>
 
 int main()
 {
-    int number_1 = 0;
-    int number_2 = 0;
+	double number_1 = 0.0;
+	double number_2 = 0.0;
 
-    std::cout << "THIS IS SIMPLE PROGRAM TO ADD TWO NUMBERS" << std::endl;
+	std::cout << "THIS IS SIMPLE PROGRAM TO ADD TWO NUMBERS" << std::endl;
 
-    std::cout << "Enter the first number : " << std::endl;
-    std::cin >> number_1;
+	std::cout << "Enter the first number : " << std::endl;
+	std::cin >> number_1;
 
-    std::cout << "Enter the second number : " << std::endl;
-    std::cin >> number_2;
+	std::cout << "Enter the second number : " << std::endl;
+	std::cin >> number_2;
 
-    int result = AddTwoNumbers(number_1, number_2);
-    std::cout << number_1 << " + " << number_2 << " = " << result << std::endl;
+	double result = AddTwoNumbers(number_1, number_2);
+	std::cout << number_1 << " + " << number_2 << " = " << result << std::endl;
 
-    return 0;
+	return 0;
 }
